Add Rankable protocol
- Add extension helpers to Array of Rankable.
- ArtistDigest and VenueDigest are now Rankable, and do not need specialized extensions on Array anymore.
